Management - Break useless inheritances: Line.
