+++ models/management/__init__.py
@@ -2,23 +2,21 @@
 
 # Import modules
 
-# Commons
-from . import line
-
 # Dependencies
 from . import productivity_day
 
+
 # Management
-from . import mgt_line
+
+from . import mgt_line      # Dep !
+
 from . import mgt_order_line
 from . import mgt_family_line
 from . import mgt_subfamily_line
 from . import mgt_day_line
 
-from . import repo     # used by mkt
 from . import management
-from . import management_controller
-from . import management_strategy
+
 
 # Patient
 from . import mgt_patient_line
