+++ models/management/mgt_subfamily_line.py
@@ -11,13 +11,13 @@ from __future__ import print_function
 from openerp import models, fields, api
 from . import mgt_vars
 
-#class SubFamilyLine(models.Model):	
 class MgtSubfamilyLine(models.Model):	
 	"""
 	Subfamily Line
 	"""	
 	_name = 'openhealth.management.sub_family.line'
 	_order = 'amount desc'
+
 
 # ----------------------------------------------------------- Relational -------
 	management_id = fields.Many2one(
