+++ models/marketing/__init__.py
@@ -1,5 +1,11 @@
 # -*- coding: utf-8 -*-
 
+# Commons
+from . import line
+
+from . import repo     # used by mkt
+
+# Dependencies
 from . import patient_line
 
 from . import medical
